requirements.txt
typer==0.12.3
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.12.0
python-dotenv==1.0.1